+++ src/main/java/com/teamtea/eclipticseasons_patch/modules/immersive_weathering/ESPatchIWChecker.java
@@ -1,0 +1,5 @@
+package com.teamtea.eclipticseasons_patch.modules.immersive_weathering;
+
+public interface ESPatchIWChecker {
+    void eclipticseasons_multimodpatch$setRecord(IW.Config.LimitRecord record);
+}